requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
astroid==2.11.7
async-timeout==4.0.2
attrs==22.1.0
black==22.6.0
blessed==1.19.1
blessings==1.7
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
docker-pycreds==0.4.0
frozenlist==1.3.0
fsspec==2022.7.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
gpustat==1.0.0rc1
grpcio==1.48.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
isort==5.10.1
joblib==1.1.0
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mido==1.2.10
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.1
nvidia-ml-py==11.495.46
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.2.2
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.2.0
platformdirs==2.5.2
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pylint==2.14.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
sentry-sdk==1.9.0
setproctitle==1.3.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
snakeviz==2.1.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchinfo==1.7.0
torchmetrics==0.9.3
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.0
types-tqdm==4.64.4
typing-extensions==4.3.0
urllib3==1.26.11
wandb==0.13.2
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
